added cash module and cash process from collect

+++ backend/app/crud/collect_detail.py
@@ -1,7 +1,8 @@
 from sqlalchemy.orm import Session
 from sqlalchemy.util.langhelpers import ellipses_string
-from ..models import CollectDetail as CollectDetailModel, Collect
+from ..models import CollectDetail as CollectDetailModel, Collect, Cash
 from ..schemas.collect_detail import CollectDetail as CollectDetailSchema, CollectDetailDelete
+from .utils import get_open_cash, add_automatic_collect
 
 
 def update_collect(db, collect_id, amount, operation):
@@ -30,6 +31,11 @@ def get_collect_detail(db: Session, collect_id: int):
 
 
 def create_collect_detail(db: Session, collect_detail: CollectDetailSchema):
+    cash: Cash = get_open_cash(db)
+    
+    if cash is None:        #There is no cash or more than one opened
+        return None
+
     db_collect_detail = CollectDetailModel(amount=collect_detail.amount,
                                            reference=collect_detail.reference,
                                            bank_id=collect_detail.bank_id,
@@ -40,6 +46,10 @@ def create_collect_detail(db: Session, collect_detail: CollectDetailSchema):
     db.commit()
     # update_invoice(db,collect_detail.invoice_id)
     update_collect(db, db_collect_detail.collect_id,db_collect_detail.amount,'+')
+    if db_collect_detail.payment_method.name == 'Efectivo':
+        concept = f"Ingreso AUTO. FACT: {db_collect_detail.collect.invoice.invoice} - Cliente: {db_collect_detail.collect.invoice.client.name}"
+        amount = db_collect_detail.amount
+        add_automatic_collect(db, concept, amount)
     return db_collect_detail
 
 
@@ -58,11 +68,20 @@ def update_collect_detail(db: Session, collect_detail: CollectDetailSchema):
 
 # TODO: Increase stock by collect.detail.qtty in product
 def delete_collect_detail(db: Session, collect_detail: CollectDetailDelete):
+    cash: Cash = get_open_cash(db)
+    
+    if cash is None:        #There is no cash or more than one opened
+        return None
+
     collect_detail_data = db.query(CollectDetailModel).filter(
         CollectDetailModel.id == collect_detail.id).first()
     if collect_detail_data is None:
         return None
     else:
+        if collect_detail_data.payment_method.name == 'Efectivo':
+            concept = f"Reverso AUTO. FACT: {collect_detail_data.collect.invoice.invoice} - Cliente: {collect_detail_data.collect.invoice.client.name}"
+            amount = collect_detail_data.amount
+            add_automatic_collect(db, concept, amount*-1)
         db.delete(collect_detail_data)
         db.commit()
         # update_invoice(db,collect_detail_data.invoice_id)
